update: rename header file for addtl clarity
