+++ WebSearcher/classifiers/__init__.py
@@ -1,3 +1,3 @@
-from .headers import ClassifyHeaderText
+from .header_text import ClassifyHeaderText
 from .main import ClassifyMain
 from .footer import ClassifyFooter
